创建数据结构
