+++ src/utils/plugin/plugin_manager.py
@@ -2,32 +2,15 @@ import importlib
 import os
 import sys
 import traceback
-from pathlib import Path
 from types import ModuleType
-from typing import Optional, Generator, Any, TypedDict
+from typing import Generator, Any
 
 import hpyculator as hpyc
 
+from ..data_structure import MetadataDict
+
 # 这样才可以导入上层包哈哈
 sys.path.append(os.path.join(sys.path[0], ".."))
-
-
-# 插件元数据数据类型
-class MetadataDict(TypedDict, total=False):
-    input_mode: str
-    id: str
-    option: str
-    version: str
-    tag: list | str
-    save_name: str
-    quantifier: str
-    output_start: str
-    output_name: str
-    author: str | list
-    help: str
-    output_end: str
-    return_mode: int
-    fullwidth_symbol: int
 
 
 class PluginManager:
